+++ R/ggbarstats.R
@@ -40,7 +40,6 @@
 #' @inherit ggpiestats return details
 #'
 #' @examples
-#' \donttest{
 #' # for reproducibility
 #' set.seed(123)
 #'
@@ -63,7 +62,6 @@
 #'   y = Hair,
 #'   counts = Freq
 #' )
-#' }
 #' @export
 
 # defining the function