+++ R/grouped_ggbetweenstats.R
@@ -25,7 +25,7 @@
 #' @inherit ggbetweenstats return details
 #'
 #' @examples
-#' \donttest{
+#'
 #' # to get reproducible results from bootstrapping
 #' set.seed(123)
 #'
@@ -55,8 +55,6 @@
 #'   ),
 #'   messages = FALSE
 #' )
-#' }
-#'
 #' @export
 
 # defining the function