+++ R/grouped_ggdotplotstats.R
@@ -21,7 +21,6 @@
 #' @inherit ggdotplotstats return details
 #'
 #' @examples
-#' \donttest{
 #' # for reproducibility
 #' set.seed(123)
 #'
@@ -43,7 +42,6 @@
 #'   ),
 #'   messages = FALSE
 #' )
-#' }
 #' @export
 
 # defining the function