+++ R/helpers_gghistostats_graphics.R
@@ -116,7 +116,6 @@ line_labeler <- line_labeller
 #'   for centrality parameter label. (Default: `2`).
 #'
 #' @examples
-#' \donttest{
 #' library(ggplot2)
 #'
 #' # creating a ploton which lines and labels are to be superposed
@@ -134,8 +133,6 @@ line_labeler <- line_labeller
 #'   y.label.position = y_label_pos,
 #'   test.value.line = TRUE
 #' )
-#' }
-#'
 #' @keywords internal
 
 # function body