+++ NEWS.md
@@ -3,14 +3,20 @@
 BUG FIXES
 
   - `ggbetweenstats` and `ggwithinstats` no longer produce incorrect label if
-    the dataframe already contains a variable named `n` (#317).
+    the dataframe already contains a variable named `n` (#317) or variables
+    with pattern `mean` (#322).
+  - `ggbetweenstats` and `ggwithinstats` mean labels respect `k` argument (#331).
 
 MINOR
 
   - `ggcoefstats` now uses `parameters::p_value` instead of `sjstats::p_value`,
-    as requested by the maintainer of that package.
+    as requested by the maintainer of that package. This might lead to
+    differences in *p*-values for `lmer` models.
   - More models supported in `ggcoefstats`: `bracl`, `brglm2`, `glmc`, `slm`,
     `wbgee`.
+  - `ggcoefstats` gains `only.significant` argument to only display display
+    stats labels for significant effects. This can be helpful when a large
+    number of regression coefficients are to be displayed in a single plot.
 
 # ggstatsplot 0.1.2